requirements.txt
httpx
python-dotenv
supabase
supabase-py
requests